+++ src/npm/index.js
@@ -1,1 +1,2 @@
-export * from 'monaco-editor-mod';
+export * from "monaco-editor-mod";
+export * from "monaco-editor-mod/esm/monacoExtras"